backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-multipart==0.0.20
numpy<2.0.0
python-dotenv==1.0.1
langgraph==0.2.45
langchain==0.3.7
langchain-huggingface==0.1.2
huggingface-hub==0.26.5
requests==2.32.3

